requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
#gunicorn==21.2.0
Flask-Talisman==1.1.0
SQLAlchemy==2.0.46
python-dotenv==1.0.1
flask_talisman==1.1.0